Allow configurable response time from smoke tests.
Transclusion can be slow. Since the smoke tests are likely hitting an uncached result (it happens right after invalidation), it's quite likely it will hit the transclusion lambda which is expected to take a bit longer.

The default will still be 1000ms for other static sites, but it can be configured per site depending on what SLOs or other expectations there are.